- moved `VideoSection` component to be a solely UI one 	- this will allow us to reuse this component in other places
